Split `Lifetime` tests into per-type files
`src/.../types/_lifetime.py` now exposes four public symbols, so the test
file grew too large. Split it following the repository convention:

    tests/types/test_lifetime.py
      -> tests/types/_lifetime/test_<thing>.py

Test bodies and assertions are unchanged. Shared time fixtures move to the
subdirectory's `conftest.py`, and test names drop redundant type prefixes.
